fix: 修复启动闪退 (颜色转换错误)
- 修复 Color(colorLong.toULong()) → Color(colorLong.toInt())
- 影响文件: CalendarView, HomeScreen, ProjectsScreen, CheckInScreen, DayCheckInPopup

版本: 1.001

+++ app/src/main/java/com/bulubulu/recordlifeitems/ui/checkin/CheckInScreen.kt
@@ -105,7 +105,7 @@ fun CheckInScreen(
                         verticalAlignment = Alignment.CenterVertically
                     ) {
                         ProjectColorCircle(
-                            color = Color(proj.color.toULong()),
+                            color = Color(proj.color.toInt()),
                             size = 40.dp
                         )
                         Spacer(modifier = Modifier.width(12.dp))
